+++ src/components/grid/GridRow.vue
@@ -1,11 +1,26 @@
 <template>
-  <div class="enk-grid-row">
+  <div :class="styling">
     <slot></slot>
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from "vue";
+import { computed, defineComponent } from "vue";
+import { Config } from "../../core/Config";
 
-export default defineComponent({});
+export default defineComponent({
+  setup() {
+    const styling = computed(() => {
+      const element = `${Config.ABBR}-grid-row`;
+
+      return {
+        [element]: true
+      };
+    });
+
+    return {
+      styling
+    };
+  }
+});
 </script>