+++ src/core/Config.ts
@@ -1,5 +1,3 @@
 export abstract class Config {
-  public static VERSION = 0.1;
-  public static NAME = "enkelt";
-  public static ABBR = "enk";
+  public static ABBR = "birdbrain";
 }