+++ src/index.ts
@@ -8,5 +8,6 @@ import Grid from "./components/grid/Grid.vue";
 import GridCol from "./components/grid/GridCol.vue";
 import GridRow from "./components/grid/GridRow.vue";
 import Icon, { icons } from "./components/icon/Icon.vue";
+import Tooltip from "./components/tooltip/Tooltip.vue";
 
-export { Alert, Button, Icon, Grid, GridRow, GridCol, Container, Badge, Avatar, AvatarGroup, icons };
+export { Alert, Button, Icon, Grid, GridRow, GridCol, Container, Badge, Avatar, AvatarGroup, Tooltip, icons };