+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+# [1.1.0-alpha.4](https://github.com/joakimbugge/birdbrain-vue/compare/v1.1.0-alpha.3...v1.1.0-alpha.4) (2021-01-30)
+
+
+### Bug Fixes
+
+* **theming:** components now always get correct prefix ([8146ed1](https://github.com/joakimbugge/birdbrain-vue/commit/8146ed17df4e36f2fc6b3a770e0a8ac267b99c6a))
+* **theming:** remove test prefix ([37e8c42](https://github.com/joakimbugge/birdbrain-vue/commit/37e8c4227eda18bc3e6c0171af7f623ed9faaea7))
+
+# [1.1.0-alpha.3](https://github.com/joakimbugge/birdbrain-vue/compare/v1.1.0-alpha.2...v1.1.0-alpha.3) (2021-01-30)
+
+
+### Features
+
+* **tooltip:** introduce tooltip component ([16856d2](https://github.com/joakimbugge/birdbrain-vue/commit/16856d2bfca0a7deee1ac0f349c6411bf3852e09))
+
+# [1.1.0-alpha.2](https://github.com/joakimbugge/birdbrain-vue/compare/v1.1.0-alpha.1...v1.1.0-alpha.2) (2021-01-29)
+
+
+### Bug Fixes
+
+* use Config.ABBR to compose class names ([dbe9603](https://github.com/joakimbugge/birdbrain-vue/commit/dbe960374d0774a3ffb747565cc352f1997e7011))
+
+
+### Features
+
+* change prefix from enk to birdbrain ([4330e77](https://github.com/joakimbugge/birdbrain-vue/commit/4330e77cb28a8228ab261ef73c253cbc22a0bdd8))
+
+# [1.1.0-alpha.1](https://github.com/joakimbugge/birdbrain-vue/compare/v1.0.0...v1.1.0-alpha.1) (2021-01-29)
+
+
+### Bug Fixes
+
+* use Config.ABBR to compose class names ([3666922](https://github.com/joakimbugge/birdbrain-vue/commit/36669221b424a1a441a65980a7ec53e0c75beac7))
+
+
+### Features
+
+* change prefix from enk to birdbrain ([7423561](https://github.com/joakimbugge/birdbrain-vue/commit/7423561c65951703af46d2cd35f12dd0b65bd4f7))
+
 # 1.0.0 (2021-01-29)